% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_shape}
\alias{make_shape}
\title{Creates a 3D shape matrix suitable}
\usage{
make_shape(shape = c("cone", "ellipsoid", "halfellipsoid", "paraboloid",
  "cylinder", "yokozawa"), nz = 25, nalpha = 25, zvals = NULL, eta = 4)
}
\arguments{
\item{shape}{Desired shape of object}

\item{nz}{Number of vertical points used to generate the shape. Only used if zvals=NULL.}

\item{nalpha}{Number of angles at which nodes in the canopy are created}

\item{zvals}{Heights to calculate points. Overrides specification of \code{nz}.}

\item{eta}{Canopy shape factor, required when shape is set as \code{yokozawa}.
See \code{yokozawa_density} for details.}
}
\value{
A matrix  with columns [x,y,z] defining the shape. These points define triangles. Note
the matrix includes duplicate points because each is used in multiple triangles.
}
\description{
Creates a 3D shape matrix suitable for passing to \code{rgl::plot.triangles}, and used
by \code{plot3d_shape}) This object has height 1, width 1, and is located at (0,0,0). Use
the function \code{resize_shape} to scale and move to a different set of coordinates
}

