% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_stand}
\alias{make_stand}
\title{Creates a stand of trees to plot in 3D}
\usage{
make_stand(trees, scaling = c(1, 1), ...)
}
\arguments{
\item{trees}{A dataframe of values for the stand. Required values are defined in code{make_stand_dataframe}}

\item{scaling}{a vector of length two, giving scaling of heights and widths in the plot}

\item{...}{Additional arguments defining crown, to pass through to \code{make_shape}}
}
\description{
Creates a stand of trees to plot in 3D
}

