% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_stand_dataframe}
\alias{make_stand_dataframe}
\title{Takes a dataframe of tree values, checks for essential variables and fills missing
values with defaults}
\usage{
make_stand_dataframe(trees, crown_shape = "yokozawa", eta = 5,
  crown_color = "forestgreen", stem_color = "brown")
}
\arguments{
\item{trees}{A dataframe of values for the stand.}

\item{crown_shape}{Shape of crown. See \code{make_shape} for allowable values.}

\item{eta}{Canopy shape factor, required when shape is set as \code{yokozawa}.
See \code{yokozawa_density} for details.}

\item{crown_color}{Colour of the crown}

\item{stem_color}{Colour of the stem}
}
\value{
A dataframe of values for the stand.
}
\description{
Takes a dataframe of tree values, checks for essential variables and fills missing
values with defaults
}

