% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_tree}
\alias{make_tree}
\title{Creates a single tree with specified characteristics}
\usage{
make_tree(x = 0, y = 0, top_height = 1, height_crown_base = 0,
  crown_width = 1, dbh = 0.01, crown = NULL, crown_shape = c("cone",
  "elipsoid", "ellipsoid", "round", "halfellipsoid", "paraboloid", "cylinder",
  "yokozawa"), eta = 13, crown_color = "forestgreen", stem = NULL,
  stem_color = "brown", nz = 25, nalpha = 25, zvals = NULL,
  scaling = c(1, 1), ...)
}
\arguments{
\item{x}{Location in x dimension}

\item{y}{Location in x dimension}

\item{top_height}{Top height of tree}

\item{height_crown_base}{Height of crown base}

\item{crown_width}{Width of crown}

\item{dbh}{Diameter at breast height}

\item{crown}{A matrix with columns [X,Y,Z] describing shape of the crown}

\item{crown_shape}{One of 'cone', 'elipsoid', 'ellipsoid', 'round', 'halfellipsoid', 'paraboloid', 'cylinder', 'yokozawa'}

\item{eta}{Canopy scaling factor. See \code{yokozawa_density} for details.}

\item{crown_color}{Colour of the crown}

\item{stem}{A matrix with columns [X,Y,Z] describing shape of the stem}

\item{stem_color}{Colour of the stem}

\item{nz}{Number of vertical points used to generate the shape. Only used if zvals=NULL.}

\item{nalpha}{Number of angles at which nodes in the canopy are created}

\item{zvals}{Heights to calculate points. Overrides specification of \code{nz}.}

\item{scaling}{a vector of length two, giving scaling of heights and widths in the plot}

\item{...}{Additional arguments defining crown, to pass through to \code{make_shape}}
}
\value{
A list with elements  "crown", "crown_color", "stem", "stem_color" defining the
shape and colour of crown and stem
}
\description{
Creates a single tree with specified characteristics. Uses existing shape matrices for \code{stem}
and \code{crown} if provided, leading
to faster computing times. Otherwise recalculates these.
}
\seealso{
\code{\link{plot_stand}}, \code{\link{plot_tree}}
}

