% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{new_rgl}
\alias{new_rgl}
\title{Creates a new rgl scene with desired dimensions}
\usage{
new_rgl(size = c(1, 1, 500, 500))
}
\arguments{
\item{size}{A vector of dimensions for the plot, as specified in \code{new_rgl}}
}
\description{
Creates a new rgl scene with desired dimensions
}

