% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.tree3d}
\alias{plot.tree3d}
\title{Plot a single tree in 3d}
\usage{
\method{plot}{tree3d}(trees, ...)
}
\arguments{
\item{...}{Additional arguments passed through to \code{rgl::rgl.triangles}}

\item{tree}{Top height of plant}
}
\description{
Plot a single tree in 3d
}

