% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.trees3d}
\alias{plot.trees3d}
\title{Plot a stand of trees}
\usage{
\method{plot}{trees3d}(stand, ...)
}
\arguments{
\item{stand}{A list of trees, as generated \code{stand3d}. elements of the list are generated
by \code{tree3d}.}

\item{...}{Additional arguments passed through to \code{rgl}}
}
\description{
Plot a stand of trees. To optimsie computaional performance when plotting via rgl,
the function groups together everything with similar colour and plots these in a single call.
}

