% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_shapes}
\alias{plot_shapes}
\title{Adds a shape node to the current scene}
\usage{
plot_shapes(shape, ...)
}
\arguments{
\item{shape}{Matrix with columns [X,Y,Z]}

\item{...}{Additional arguments passed through to \code{rgl::rgl.triangles}}
}
\description{
Adds a shape node to the current scene
}

