% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_stand}
\alias{plot_stand}
\title{Plot a stand of trees}
\usage{
plot_stand(stand, add = FALSE, size = c(1, 1, 500, 500), ...)
}
\arguments{
\item{stand}{A list containing trees, as generated \code{make_stand}}

\item{add}{Add to existing plot?}

\item{size}{A vector of dimensions for the plot, as specified in \code{new_rgl}}

\item{...}{Additional arguments passed through to \code{rgl::rgl.triangles}}
}
\description{
Plot a stand of trees. To optimsie computaional performance when plotting via rgl,
the function groups together everything with similar colour and plots these in a single call.
}

