% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot a single tree in 3D}
\usage{
plot_tree(tree, ...)
}
\arguments{
\item{tree}{Top height of plant}

\item{...}{Additional arguments passed through to \code{rgl::rgl.triangles}}
}
\description{
Plot a single tree in 3D
}

