% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{resize_shape}
\alias{resize_shape}
\title{Scale and reposition a shape}
\usage{
resize_shape(shape, height = 1, width = 1, x = 0, y = 0, z = 0)
}
\arguments{
\item{shape}{Initial shape, defined by a matrix with columns [X,Y,Z]}

\item{height}{Height of resized object}

\item{width}{Width of resized object}

\item{x}{Offset in X dimension}

\item{y}{Offset in Y dimension}

\item{z}{Offset in Z dimension}
}
\value{
Resized and repositioned shape, defined by a matrix with columns [X,Y,Z]
}
\description{
Scale and reposition a shape.
}

