% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stand3d}
\alias{stand3d}
\title{Creates a stand of trees to plot in 3d}
\usage{
stand3d(trees, ...)
}
\arguments{
\item{trees}{A dataframe of values for the stand. Required values are defined in code{stand3d_dataframe}}

\item{...}{Additional arguments defining crown, to pass through to \code{tree_shape3d}}
}
\value{
Returns a list of trees in the stand
}
\description{
Creates a stand of trees to plot in 3d
}

