% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tree3d}
\alias{tree3d}
\title{Creates a single tree with specified characteristics}
\usage{
tree3d(x = 0, y = 0, height = 1, height_crown_base = 0.5,
  crown_width = 1, dbh = 0.01, crown_shape = c("cone", "ellipsoid",
  "round", "halfellipsoid", "paraboloid", "cylinder", "yokozawa"), eta = 13,
  crown_color = "forestgreen", crown = NULL, stem = NULL,
  stem_color = "brown", nz = 25, nalpha = 25, zvals = NULL, ...)
}
\arguments{
\item{x}{Location in x dimension}

\item{y}{Location in x dimension}

\item{height}{Top height of tree}

\item{height_crown_base}{Height of crown base}

\item{crown_width}{Width of crown}

\item{dbh}{Diameter at breast height}

\item{crown_shape}{One of 'cone', 'ellipsoid', 'round', 'halfellipsoid', 'paraboloid', 'cylinder', 'yokozawa'}

\item{eta}{Canopy scaling factor. See \code{yokozawa_density} for details.}

\item{crown_color}{Colour of the crown}

\item{crown}{A previously existing matrix with columns [X,Y,Z] describing shape of the crown}

\item{stem}{A previously existing matrix with columns [X,Y,Z] describing shape of the stem}

\item{stem_color}{Colour of the stem}

\item{nz}{Number of vertical points used to generate the shape. Only used if zvals=NULL.}

\item{nalpha}{Number of angles at which nodes in the canopy are created}

\item{zvals}{Heights to calculate points. Overrides specification of \code{nz}.}

\item{...}{Additional arguments defining crown, to pass through to \code{tree_shape3d}}
}
\value{
A list with elements  "crown" & "stem", ecah of which has elements "shape" and "col"
}
\description{
Creates a single tree with specified characteristics. Uses existing shape matrices for \code{stem}
and \code{crown} if provided, leading
to faster computing times. Otherwise recalculates these.
}
\seealso{
\code{\link{plot.tree3d}}
}

