% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tree_shape3d}
\alias{tree_shape3d}
\title{Creates a 3d shape matrix suitable for plotting with rgl devices}
\usage{
tree_shape3d(shape = c("cone", "ellipsoid", "halfellipsoid", "paraboloid",
  "cylinder", "yokozawa"), nz = 25, nalpha = 25, zvals = NULL, eta = 4)
}
\arguments{
\item{shape}{Desired shape of object: "cone", "ellipsoid", "halfellipsoid",
"paraboloid", "cylinder", or "yokozawa".}

\item{nz}{Number of vertical points used to generate the shape.}

\item{nalpha}{Number of angles at which nodes in the canopy are created}

\item{zvals}{Heights to calculate points. Overrides specification of \code{nz}.}

\item{eta}{Canopy shape factor, required when shape is set as \code{yokozawa}.
See \code{yokozawa_density} for details.}
}
\value{
A matrix  with columns [x,y,z] defining the shape. These points define triangles. Note
the matrix includes duplicate points because each is used in multiple triangles.
}
\description{
Creates a 3d shape matrix suitable for passing to \code{rgl} plotting devices,
such as \code{triangles3d}. This object has height 1, width 1, and is located at (0,0,0). Use
the functions \code{rotate3d, scale3d, translate3d} in \code{rgl} to scale and move to a different set of coordinates
}

