% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{yokozawa_density}
\alias{yokozawa_density}
\title{Function describing vertical distribution of leaf area within the crown}
\usage{
yokozawa_density(z, h, eta)
}
\arguments{
\item{z}{Height within canopy of the plant}

\item{h}{Top height of plant}

\item{eta}{Canopy shape factor}
}
\value{
A vector of values, with same length as z
}
\description{
Function describing vertical distribution of leaf area within the crown of a plant.
First described by Yokozawa 1995 doi: 10.1006/anbo.1995.1096
}

